chore: deprecate numberFormat
