+++ utils/format.js
@@ -1,6 +1,6 @@
-const numberFormat = require('./numberFormat');
 const luxon = require('luxon');
 const frappe = require('frappejs');
+const { DEFAULT_DISPLAY_PRECISION, DEFAULT_LOCALE } = require('./consts');
 
 module.exports = {
   format(value, df, doc) {
@@ -13,7 +13,8 @@ module.exports = {
     }
 
     if (df.fieldtype === 'Currency') {
-      value = formatCurrency(value, df, doc);
+      const currency = getCurrency(df, doc);
+      value = formatCurrency(value, currency);
     } else if (df.fieldtype === 'Date') {
       let dateFormat;
       if (!frappe.SystemSettings) {
@@ -47,7 +48,61 @@ module.exports = {
     }
     return value;
   },
+  formatCurrency,
+  formatNumber,
 };
+
+function formatCurrency(value, currency) {
+  let valueString;
+  try {
+    valueString = formatNumber(value);
+  } catch (err) {
+    err.message += ` value: '${value}', type: ${typeof value}`;
+    throw err;
+  }
+
+  const currencySymbol = frappe.currencySymbols[currency];
+  if (currencySymbol) {
+    return currencySymbol + ' ' + valueString;
+  }
+
+  return valueString;
+}
+
+function formatNumber(value) {
+  const currencyFormatter = getNumberFormatter();
+  if (typeof value === 'number') {
+    return currencyFormatter.format(value);
+  }
+
+  if (value.round) {
+    return currencyFormatter.format(value.round());
+  }
+
+  const formattedCurrency = currencyFormatter.format(value);
+  if (formattedCurrency === 'NaN') {
+    throw Error(
+      `invalid value passed to formatCurrency: '${value}' of type ${typeof value}`
+    );
+  }
+
+  return formattedCurrency;
+}
+
+function getNumberFormatter() {
+  if (frappe.currencyFormatter) {
+    return frappe.currencyFormatter;
+  }
+
+  const locale = frappe.SystemSettings.locale ?? DEFAULT_LOCALE;
+  const display =
+    frappe.SystemSettings.displayPrecision ?? DEFAULT_DISPLAY_PRECISION;
+
+  return (frappe.currencyFormatter = Intl.NumberFormat(locale, {
+    style: 'decimal',
+    minimumFractionDigits: display,
+  }));
+}
 
 function getCurrency(df, doc) {
   if (!(doc && df.getCurrency)) {
@@ -59,22 +114,4 @@ function getCurrency(df, doc) {
   }
 
   return df.getCurrency(doc);
-}
-
-function formatCurrency(value, df, doc) {
-  const currency = getCurrency(df, doc);
-  let valueString;
-  try {
-    valueString = numberFormat.formatCurrency(value);
-  } catch (err) {
-    err.message += ` value: '${value}', type: ${typeof value}`;
-    console.error(df);
-    throw err;
-  }
-  const currencySymbol = frappe.currencySymbols[currency];
-
-  if (currencySymbol) {
-    return currencySymbol + ' ' + valueString;
-  }
-  return valueString;
 }
